pyproject.toml
[project]
name = "teamchallenge-1444"
version = "0.1.0"
description = "backend"
authors = [
    {name = "Rostyslav Bysko",email = "brn.h74@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "attrs (>=25.1.0,<26.0.0)",
    "django-filter (>=24.3,<25.0)",
    "django-jsoneditor (>=0.2.4,<0.3.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "drf-nested-routers (>=0.94.1,<0.95.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "inflection (>=0.5.1,<0.6.0)",
    "jsonschema (>=4.23.0,<5.0.0)",
    "packaging (>=24.2,<25.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "referencing (>=0.36.2,<0.37.0)",
    "rpds-py (>=0.22.3,<0.23.0)",
    "djoser (>=2.3.1,<3.0.0)",
    "django-model-utils (>=5.0.0,<6.0.0)",
    "pip (>=25.0,<26.0)",
    "django-treebeard (>=4.7.1,<5.0.0)"

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
